refactor(DEV-001): Split database schemas into individual files
- Created individual schema files in packages/backend/src/schemas/:
  - user.schema.ts (userRoleEnum, userStatusEnum, users)
  - passwordReset.schema.ts (passwordResetTokens)
  - conversation.schema.ts (conversationStatusEnum, conversationPriorityEnum, channelTypeEnum, conversations)
  - message.schema.ts (messageStatusEnum, messageDirectionEnum, messages)
  - tag.schema.ts (tags, conversationTags)
  - note.schema.ts (notes)
  - notification.schema.ts (notifications)
  - routingRule.schema.ts (routingRules, routingRuleExecutions)
  - auditLog.schema.ts (auditLogs)
  - rawPayload.schema.ts (rawPayloads)
  - attachment.schema.ts (attachments)
  - betterAuth.schema.ts (session, verification, account)
  - index.ts (barrel exports)

- Updated db.schema.ts to re-export all schemas from individual files
- Follows ADR-005: One definition per file principle
- Enables clearer separation of concerns and easier maintenance
