+++ packages/backend/src/schemas/index.ts
@@ -1,0 +1,23 @@
+// User & Auth
+export * from './user.schema';
+export * from './passwordReset.schema';
+export * from './betterAuth.schema';
+
+// Conversations & Messages
+export * from './conversation.schema';
+export * from './message.schema';
+
+// Collaboration
+export * from './tag.schema';
+export * from './note.schema';
+
+// Notifications
+export * from './notification.schema';
+
+// Routing & Rules
+export * from './routingRule.schema';
+
+// Audit & Payloads
+export * from './auditLog.schema';
+export * from './rawPayload.schema';
+export * from './attachment.schema';
