+++ packages/backend/src/schemas/notification.schema.ts
@@ -1,0 +1,33 @@
+import { pgTable, uuid, varchar, text, timestamp, boolean, jsonb, index } from 'drizzle-orm/pg-core';
+import { users } from './user.schema';
+import { conversations } from './conversation.schema';
+
+/**
+ * Notifications table - in-app notifications for users
+ */
+export const notifications = pgTable(
+  'notifications',
+  {
+    id: uuid('id').primaryKey().defaultRandom(),
+    userId: uuid('user_id')
+      .notNull()
+      .references(() => users.id, { onDelete: 'cascade' }),
+    type: varchar('type', { length: 50 }).notNull(), // 'assignment', 'mention', etc.
+    conversationId: uuid('conversation_id').references(() => conversations.id, {
+      onDelete: 'set null',
+    }),
+    actorId: uuid('actor_id').references(() => users.id, { onDelete: 'set null' }),
+    message: text('message').notNull(),
+    isRead: boolean('is_read').notNull().default(false),
+    metadata: jsonb('metadata'),
+    createdAt: timestamp('created_at').notNull().defaultNow(),
+  },
+  (table) => ({
+    userIdIdx: index('notifications_user_id_idx').on(table.userId),
+    isReadIdx: index('notifications_is_read_idx').on(table.isRead),
+    createdAtIdx: index('notifications_created_at_idx').on(table.createdAt),
+  })
+);
+
+export type Notification = typeof notifications.$inferSelect;
+export type NotificationInsert = typeof notifications.$inferInsert;
