+++ packages/backend/src/schemas/passwordReset.schema.ts
@@ -1,0 +1,27 @@
+import { pgTable, serial, timestamp, varchar, index, uuid } from 'drizzle-orm/pg-core';
+import { users } from './user.schema';
+
+/**
+ * PasswordResetTokens table - stores password reset tokens with expiry
+ */
+export const passwordResetTokens = pgTable(
+  'password_reset_tokens',
+  {
+    id: serial('id').primaryKey(),
+    userId: uuid('user_id')
+      .notNull()
+      .references(() => users.id, { onDelete: 'cascade' }),
+    token: varchar('token', { length: 255 }).notNull().unique(),
+    expiresAt: timestamp('expires_at').notNull(),
+    usedAt: timestamp('used_at'),
+    createdAt: timestamp('created_at').notNull().defaultNow(),
+  },
+  (table) => ({
+    userIdIdx: index('password_reset_tokens_user_id_idx').on(table.userId),
+    tokenIdx: index('password_reset_tokens_token_idx').on(table.token),
+    expiresAtIdx: index('password_reset_tokens_expires_at_idx').on(table.expiresAt),
+  })
+);
+
+export type PasswordResetToken = typeof passwordResetTokens.$inferSelect;
+export type PasswordResetTokenInsert = typeof passwordResetTokens.$inferInsert;
