+++ packages/backend/src/schemas/user.schema.ts
@@ -1,0 +1,33 @@
+import {boolean, pgTable, text, timestamp, uuid, varchar, index, pgEnum} from "drizzle-orm/pg-core";
+
+export const userRoleEnum = pgEnum('user_role', [
+  'super_admin',
+  'admin',
+  'manager',
+  'user',
+]);
+
+export const userStatusEnum = pgEnum('user_status', ['active', 'inactive', 'suspended']);
+
+
+export const users = pgTable(
+  'users',
+  {
+    id: uuid('id').primaryKey().defaultRandom(),
+    email: varchar('email', { length: 255 }).notNull().unique(),
+    name: varchar('name', { length: 255 }).notNull(),
+    passwordHash: text('password_hash').notNull(),
+    role: userRoleEnum('role').notNull().default('user'),
+    status: userStatusEnum('status').notNull().default('active'),
+    emailVerified: boolean('email_verified').notNull().default(false),
+    image: text('image'),
+    createdAt: timestamp('created_at').notNull().defaultNow(),
+    updatedAt: timestamp('updated_at').notNull().defaultNow(),
+    lastLoginAt: timestamp('last_login_at'),
+  },
+  (table) => ({
+    emailIdx: index('users_email_idx').on(table.email),
+    roleIdx: index('users_role_idx').on(table.role),
+    statusIdx: index('users_status_idx').on(table.status),
+  })
+);
